requirements.txt
# 核心数据处理库
pandas>=2.2.0
numpy>=1.24.0
scipy>=1.11.0
requests>=2.31.0

# 数据获取
yfinance>=0.2.28
beautifulsoup4>=4.12.0
lxml>=4.9.0
alpha-vantage>=2.3.1
quandl>=3.7.0

# 数据库
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
pymongo>=4.2.0

# 机器学习
scikit-learn>=1.3.0
statsmodels>=0.14.0
xgboost>=1.6.0
lightgbm>=3.3.0

# 技术分析 (使用兼容版本)
pandas-ta>=0.4.67b0
TA-Lib>=0.4.25

# 优化算法
cvxpy>=1.2.0
cvxopt>=1.3.0

# 可视化
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
bokeh>=2.4.0

# Web框架
FastAPI>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# 配置管理
python-dotenv>=1.0.0
PyYAML>=6.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Web相关
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# 浏览器自动化
selenium>=4.15.0
webdriver-manager>=4.0.0

# 系统监控
psutil>=5.9.0
memory-profiler>=0.61.0
loguru>=0.7.0

# 异步处理
aiohttp>=3.9.0
asyncio
celery>=5.3.0

# 安全
cryptography>=41.0.0
bcrypt>=4.1.0

# 时间处理
pytz>=2023.3
python-dateutil>=2.8.2

# 网络和API
httpx>=0.25.0
websocket-client>=1.6.0
urllib3>=2.0.0

# 数据验证
pydantic>=2.5.0
marshmallow>=3.20.0

# 缓存
cachetools>=5.3.0

# 数学计算
sympy>=1.12.0

# 开发工具
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# 文档生成
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# 部署
gunicorn>=21.2.0
supervisor>=4.2.0

# 监控
prometheus-client>=0.19.0

# 邮件通知
email-validator>=2.1.0

# 文件处理
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# API限流
ratelimit>=2.2.1
backoff>=2.2.1

# 环境变量
python-decouple>=3.8

# JSON处理
ujson>=5.8.0

# 加密货币支持（可选）
ccxt>=4.1.0

# 风险管理（可选）
# quantlib>=1.32
pytest-cov>=4.1.0
pytest-asyncio>=0.19.0

# 代码质量
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=0.971

# 日志
loguru>=0.7.0

# 数据验证
pydantic>=1.10.0

# 性能分析
memory-profiler>=0.60.0
line-profiler>=3.5.0

# 数学计算
arch>=5.3.0

# 时间处理
python-dateutil>=2.8.0
pytz>=2022.1

# 缓存
diskcache>=5.4.0

# 进度条
tqdm>=4.64.0

# 数据格式
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# 加密
cryptography>=37.0.0

# 环境变量
python-decouple>=3.6

# 异步处理
aiohttp>=3.8.0

# 工具库
tqdm>=4.66.0
joblib>=1.3.0

# 时间处理
pytz>=2023.3
python-dateutil>=2.8.0

# 开发工具
ipython>=8.17.0
jupyter>=1.0.0

# 数据验证
pydantic>=2.5.0

# 缓存
diskcache>=5.6.0

# 异步支持
aiohttp>=3.9.0
aiofiles>=23.2.0

# 高级量化分析库 (Python 3.12 兼容)
TA-Lib>=0.4.25  # 技术分析库
numba>=0.58.0  # 高性能数值计算
empyrical-reloaded>=0.5.0  # 金融风险和性能指标（兼容Python 3.12的维护版）

# 量化回测框架
zipline-reloaded>=3.0.0  # 量化回测引擎
pyfolio-reloaded>=0.9.3  # 投资组合分析（兼容维护版）

# 投资组合优化
riskfolio-lib>=5.1.0  # 现代投资组合理论

# 金融数学库
QuantLib>=1.32  # 金融衍生品定价（正确的PyPI包名）

# 新增因子计算器依赖包
# 深度学习框架
torch>=2.0.0  # PyTorch深度学习框架
torchvision>=0.15.0  # 计算机视觉库
tensorflow>=2.13.0  # TensorFlow深度学习框架（可选）

# 高级机器学习
catboost>=1.2.0  # CatBoost梯度提升
imbalanced-learn>=0.11.0  # 不平衡数据处理
hyperopt>=0.2.7  # 超参数优化
optuna>=3.4.0  # 自动机器学习优化

# 时间序列分析
statsforecast>=1.6.0  # 统计时间序列预测
prophet>=1.1.4  # Facebook时间序列预测
arch>=6.2.0  # ARCH/GARCH模型

# 网络分析
networkx>=3.2.0  # 网络图分析
igraph>=0.10.0  # 高性能图分析

# 信号处理
PyWavelets>=1.4.0  # 小波变换
spectrum>=0.8.1  # 频谱分析

# 高级数值计算
numexpr>=2.8.0  # 快速数值表达式计算
bottleneck>=1.3.0  # 快速NumPy数组函数

# 统计分析
pingouin>=0.5.3  # 统计分析库
factor-analyzer>=0.4.1  # 因子分析

# 自然语言处理（情绪分析）
textblob>=0.17.1  # 简单文本处理
vaderSentiment>=3.3.2  # 情绪分析
transformers>=4.35.0  # Hugging Face变换器（可选）

# 经济数据获取
fredapi>=0.5.1  # 美联储经济数据
pandas-datareader>=0.10.0  # 多源数据获取

# 券商API
alpaca-trade-api>=3.1.1  # Alpaca Markets API

# 期权定价
py_vollib>=1.0.1  # 期权波动率库
mibian>=0.1.3  # Black-Scholes期权定价

# 高频数据处理
arctic>=1.79.4  # 高性能时间序列数据库
pystore>=0.1.23  # 快速数据存储

# 风险模型
empyrical>=0.5.5  # 风险指标计算
pyportfolioopt>=1.5.5  # 投资组合优化

# 另类数据处理
tweepy>=4.14.0  # Twitter API（社交媒体数据）
praw>=7.7.1  # Reddit API（社交媒体数据）
newspaper3k>=0.2.8  # 新闻文章抓取

# 宏观经济指标
wbgapi>=1.0.12  # 世界银行数据
oecd>=0.2.5  # OECD数据

# 加密货币数据
python-binance>=1.0.19  # 币安API
cryptocompare>=0.7.6  # 加密货币数据

# 高性能计算
dask>=2023.11.0  # 并行计算
ray>=2.8.0  # 分布式计算（可选）

# 数据质量检查
great-expectations>=0.18.0  # 数据质量验证
pandera>=0.17.0  # 数据验证

# 可视化增强
plotly-dash>=2.14.0  # 交互式仪表板
altair>=5.1.0  # 声明式可视化